Refactor instructions to store individual pieces of data instead of just using an offset into a buffer
Also included is a partial implementation of the logic needed to automatically fix various types of issues

git-svn-id: https://smali.googlecode.com/svn/trunk@502 55b6fa8a-2a1e-11de-a435-ffa8d773f76a

+++ dexlib/src/main/java/org/jf/dexlib/Util/ByteArrayInput.java
@@ -111,7 +111,7 @@ public class ByteArrayInput
         }
 
         cursor = end;
-        return (int)((data[readAt] & 0xff) +
+        return ((data[readAt] & 0xff) +
                 ((data[readAt + 1] & 0xff) << 8));
     }
